Change Crude Protein from int to decimal

+++ Server/src/SERVERAPI/Models/FarmData.cs
@@ -78,7 +78,7 @@ namespace SERVERAPI.Models
         public decimal remN { get; set; }
         public decimal remP2o5 { get; set; }
         public decimal remK2o { get; set; }
-        public int crudeProtien { get; set; }
+        public decimal crudeProtien { get; set; }
         public int prevCropId { get; set; }
         public bool? coverCropHarvested { get; set; }
     }
